requirements.txt
django
django-configurations 
dj-database-url
crispy-bootstrap5

